Extract PPM closeout tests to seperate file and do light refactoring
